🐛 fix anniversary blocking notifications
