+++ src/mainpageCore.ts
@@ -1,3 +1,4 @@
+import anniversary from './components/anniversary.vue';
 import he from 'he';
 import LSSMMenu from './LSSM-Menu.vue';
 import telemetry from './modules/telemetry/main';
@@ -95,6 +96,32 @@ export default async (LSSM: Vue): Promise<void> => {
                 i18n: LSSM.$i18n,
                 render: h => h(LSSMMenu),
             }).$mount(indicatorWrapper);
+
+            if (
+                new Date() >= new Date('2021-11-20T00:00') &&
+                new Date() < new Date('2021-11-29T00:00')
+            ) {
+                LSSM.$store
+                    .dispatch('settings/getSetting', {
+                        moduleId: 'global',
+                        settingId: 'anniversary1Clicked',
+                        defaultValue: false,
+                    })
+                    .then((clicked: boolean) => {
+                        if (!clicked) {
+                            const anniversaryWrapper = document.createElement(
+                                'div'
+                            );
+                            anniversaryWrapper.id = 'aniwrap';
+                            document.body.append(anniversaryWrapper);
+                            new LSSM.$vue({
+                                store: LSSM.$store,
+                                i18n: LSSM.$i18n,
+                                render: h => h(anniversary),
+                            }).$mount(anniversaryWrapper);
+                        }
+                    });
+            }
 
             LSSM.$store
                 .dispatch('settings/getSetting', {
