+++ src/store/api.ts
@@ -201,7 +201,7 @@ const updateVehicleStates = (
     commit(`${commitFromRoot ? 'api/' : ''}setVehicleStates`, states);
 };
 
-let vehicleStorageUpdateTimeout = 0;
+const vehicleStorageUpdateTimeout = 0;
 
 export default {
     namespaced: true,
@@ -302,22 +302,22 @@ export default {
             }
             if (vehicleStorageUpdateTimeout)
                 window.clearTimeout(vehicleStorageUpdateTimeout);
-            vehicleStorageUpdateTimeout = window.setTimeout(
-                () =>
-                    set_api_storage(
-                        'vehicles',
-                        {
-                            value: state.vehicles,
-                            lastUpdate:
-                                state.lastUpdates.vehicles ??
-                                new Date().getTime(),
-                            user_id: window.user_id,
-                        },
-                        (this as unknown) as Store<RootState>,
-                        true
-                    ),
-                1000
-            );
+            // vehicleStorageUpdateTimeout = window.setTimeout(
+            //     () =>
+            //         set_api_storage(
+            //             'vehicles',
+            //             {
+            //                 value: state.vehicles,
+            //                 lastUpdate:
+            //                     state.lastUpdates.vehicles ??
+            //                     new Date().getTime(),
+            //                 user_id: window.user_id,
+            //             },
+            //             (this as unknown) as Store<RootState>,
+            //             true
+            //         ),
+            //     1000
+            // );
         },
         enableAutoUpdate(state: APIState, api: StorageAPIKey) {
             state.autoUpdates.push(api);
