+++ typings/LSSMV4.d.ts
@@ -1,9 +1,7 @@
-import { IconDefinition } from '@fortawesome/fontawesome-svg-core';
 import { NotificationsState } from './store/notifications/State';
 
 export interface LSSMV4Data {
     id: string;
-    faTimes: IconDefinition;
 }
 
 export interface LSSMV4Computed {
